Extract optionValue into a module
